LayoutTests/fast/gradients/css3-gradient-parsing-expected.txt

Tests -webkit-linear-gradient and -webkit-radial-gradient

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testGradient("background-image: -webkit-linear-gradient(black 0%", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(top)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(10deg)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(top center, black 0%)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(10px 20%, black 0%)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(left left)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(left center)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(left 10deg, black 0%)", "background-image") is "none"
PASS testGradient("background-image: -webkit-linear-gradient(black 0)", "background-image") is "-webkit-linear-gradient(top, black 0px)"
PASS testGradient("background-image: -webkit-linear-gradient(black 0%)", "background-image") is "-webkit-linear-gradient(top, black 0%)"
PASS testGradient("background-image: -webkit-linear-gradient(black 0%)", "background-image") is "-webkit-linear-gradient(top, black 0%)"
PASS testGradient("background-image: -webkit-linear-gradient(black, white)", "background-image") is "-webkit-linear-gradient(top, black, white)"
PASS testGradient("background-image: -webkit-linear-gradient(black 10px, white 20px)", "background-image") is "-webkit-linear-gradient(top, black 10px, white 20px)"
PASS testGradient("background-image: -webkit-linear-gradient(top left, black 0%)", "background-image") is "-webkit-linear-gradient(left top, black 0%)"
PASS testGradient("background-image: -webkit-linear-gradient(top, black 0%)", "background-image") is "-webkit-linear-gradient(top, black 0%)"
PASS testGradient("background-image: -webkit-linear-gradient(10deg, black 0%)", "background-image") is "-webkit-linear-gradient(10deg, black 0%)"
PASS successfullyParsed is true

TEST COMPLETE
